Replaced original file with temporary file
